// src/app/models/3dModel/imodel/ibuilding-geometry.ts
export class IBuildingGeometry {
  BuildingNumber: number;
  BuildingName: string;
  BuildingType: string;
  FrameType: string;
  ElevationA: string;
  Width: number;
  Length: number;
  EaveHeight1: number;
  EaveHeight2: number;
  DistToRidge1: number;
  DistToRidge2: number;
  RoofSlope1: number;
  RoofSlope2: number;
  PeakHeight: number;
  BaseElevation: number;
  //AttachedToBuildingNumber: number;
  //AttachedToWall: string;
  AttachmentWall: string;
  XDirection: string;
  XDistance: number;
  YDirection: string;
  YDistance: number;
  AttachmentOffset: number;
  XCoordinate: number;
  YCoordinate: number;
  BuildingLabel: string;
  //ProjectNumber: number;
  //Existing: boolean;
  HighSideWall: string;
}